Change int custom file to long as pipedrive allows numbers bigger than int32

+++ src/Pipedrive.net/Models/Common/CustomFields/IntCustomField.cs
@@ -3,9 +3,9 @@
     // int
     public class IntCustomField : ICustomField
     {
-        public int Value { get; set; }
+        public long Value { get; set; }
 
-        public IntCustomField(int value)
+        public IntCustomField(long value)
         {
             Value = value;
         }
